+++ .gitignore
@@ -2,3 +2,6 @@
 py2
 py3
 venv
+build/*
+dist/*
+src/sqlitefid.egg-info/*
