+++ pyproject.toml
@@ -1,0 +1,3 @@
+[build-system]
+requires = ["setuptools>=43.0.0", "wheel"]
+build-backend = "setuptools.build_meta"
