+++ setup.cfg
@@ -1,0 +1,2 @@
+[metadata]
+license_files = LICENSE
