+++ src/sqlitefid/libs/SFLoaderClass.py
@@ -11,12 +11,9 @@ into the sqlite db.
 
 from __future__ import absolute_import
 
-if __name__.startswith("sqlitefid"):
-    from sqlitefid.libs.SFHandlerClass import SFYAMLHandler
-    from sqlitefid.libs.ToolMappingClass import ToolMapping
-else:
-    from libs.SFHandlerClass import SFYAMLHandler
-    from libs.ToolMappingClass import ToolMapping
+# Required imports for main from root.
+from .SFHandlerClass import SFYAMLHandler
+from .ToolMappingClass import ToolMapping
 
 
 class SFLoader:
