+++ src/sqlitefid/libs/ToolMappingClass.py
