+++ tests/test_identify_export.py
@@ -1,8 +1,12 @@
 # -*- coding: utf-8 -*-
 
-from __future__ import absolute_import
-
-from sqlitefid.libs.IdentifyExportClass import IdentifyExport
+try:
+    from sqlitefid.src.sqlitefid.libs.IdentifyExportClass import IdentifyExport
+except ModuleNotFoundError:
+    # Needed when imported as submodule via demystify.
+    from src.demystify.sqlitefid.src.sqlitefid.libs.IdentifyExportClass import (
+        IdentifyExport,
+    )
 
 DROID_CSV = u""""ID","PARENT_ID","URI","FILE_PATH","NAME","METHOD","STATUS","SIZE","TYPE","EXT","LAST_MODIFIED","EXTENSION_MISMATCH","SHA1_HASH","FORMAT_COUNT","PUID","MIME_TYPE","FORMAT_NAME","FORMAT_VERSION"
 "2","0","file:////10.1.4.222/gda/archives-sample-files/opf-format-corpus/format-corpus/","\\10.1.4.222\\gda\archives-sample-files\\opf-format-corpus\format-corpus","🖤format-corpus",,"Done","","Folder",,"2014-02-28T15:49:11","false",,"",,"","",""
