+++ tests/test_sf_internal_functions.py
@@ -4,7 +4,11 @@ from __future__ import absolute_import
 
 import pytest
 
-from sqlitefid.libs import PyDateHandler, SFHandlerClass
+try:
+    from sqlitefid.src.sqlitefid.libs import PyDateHandler, SFHandlerClass
+except ModuleNotFoundError:
+    # Needed when imported as submodule via demystify.
+    from src.demystify.sqlitefid.src.sqlitefid.libs import PyDateHandler, SFHandlerClass
 
 
 @pytest.mark.parametrize(
