+++ tests/test_sqlite_basedb.py
@@ -3,13 +3,16 @@
 from __future__ import absolute_import
 
 import collections
-import sys
 
 import pytest
 
-from sqlitefid.libs.GenerateBaselineDBClass import GenerateBaselineDB
-
-PY3 = bool(sys.version_info[0] == 3)
+try:
+    from sqlitefid.src.sqlitefid.libs.GenerateBaselineDBClass import GenerateBaselineDB
+except ModuleNotFoundError:
+    # Needed when imported as submodule via demystify.
+    from src.demystify.sqlitefid.src.sqlitefid.libs.GenerateBaselineDBClass import (
+        GenerateBaselineDB,
+    )
 
 
 FIDDatabase = collections.namedtuple("FIDDatabase", "baseline cursor")
@@ -22,9 +25,8 @@ def database():
     :returns: Yielded named tuple containing a baseline db object
         (GenerateBaselineDB) and database cursor object (sqlite3.Cursor)
     """
-    basedb = GenerateBaselineDB("export.csv")
+    basedb = GenerateBaselineDB("export.csv", in_memory=True)
     basedb.tooltype = "droid"
-    basedb.dbname = "file::memory:?cache=shared"
     connection = FIDDatabase(basedb, basedb.dbsetup())
     yield connection
