+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py27, py39, linting
+envlist = py39,linting
 skipsdist = True
 
 [testenv]
